fix(instrumentation): do not export Node.js types for Browser users
